refactor: resourcesを整理した
