+++ lib/resources/doc.go
@@ -1,0 +1,4 @@
+// Resourceはオブジェクトに紐付かないゲーム中の機能の役割を果たす要素である。
+// 移動の定数などは、明らかにオブジェクトに紐付かないものである。
+
+package resources
