--- lib/resources/loader.go
@@ -1,10 +1,0 @@
-package resources
-
-const (
-	exteriorSpriteNumber   = 0
-	wallSpriteNumber       = 1
-	floorSpriteNumber      = 2
-	playerSpriteNumber     = 3
-	warpNextSpriteNumber   = 4
-	warpEscapeSpriteNumber = 5
-)
